[3단계-개선-03] 패키지 구조를 도메인 별로 변경
