+++ src/main/java/kr/codesqaud/cafe/question/controller/QuestionController.java
@@ -1,0 +1,78 @@
+package kr.codesqaud.cafe.question.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+
+import kr.codesqaud.cafe.question.dto.response.QuestionWriteDTO;
+import kr.codesqaud.cafe.question.service.QuestionService;
+
+@Controller
+@RequestMapping("/questions")
+public class QuestionController {
+
+	private final QuestionService service;
+
+	public QuestionController(QuestionService service) {
+		this.service = service;
+	}
+
+	/**
+	 * Q&A 게시글 작성 페이지로 이동
+	 * @return Q&A 게시글 작성 페이지
+	 */
+	@GetMapping("/write-form")
+	public String writeForm() {
+		return "qna/form";
+	}
+
+	/**
+	 * Q&A 게시글 쓰기 기능
+	 * @param dto 게시글 쓰기용 dto
+	 * @return Q&A 게시글 작성 페이지로 redirect
+	 */
+	@PostMapping
+	public String questionAdd(QuestionWriteDTO dto) {
+		service.addQuestion(dto);
+		return "redirect:questions/write-form";
+	}
+
+	/**
+	 * Q&A 게시글 목록 페이지로 이동
+	 * @param page 페이지 번호
+	 * @param model `게시글`, `페이징 정보`를 담고 있는 dto 를 전달하기 위한 model
+	 * @return Q&A 게시글 목록 페이지
+	 */
+	@GetMapping
+	public String questionList(@RequestParam(value = "page", required = false, defaultValue = "1") int page,
+		Model model) {
+
+		model.addAttribute("questionBoardDto", service.makeQuestionBoard(page));
+
+		return "index";
+	}
+
+	/**
+	 * Q&A 게시글 상세 보기 페이지로 이동
+	 * @param questionIdx 조회하고자 하는 Q&A 게시글의 idx
+	 * @param errorMessage 없는 게시글 또는 잘못된 입력값이 들어왔을때 받아올 에러 메시지
+	 * @param model `Q&A 게시글 상세 내역` 또는 `에러 메시지`를 전달하기 위한 model
+	 * @return Q&A 게시글 상세 보기 페이지
+	 */
+	@GetMapping("/{questionIdx}")
+	public String questionDetails(@PathVariable String questionIdx, @ModelAttribute("errorMessage") String errorMessage,
+		Model model) {
+		if (errorMessage.isBlank()) {
+			int idx = Integer.parseInt(questionIdx);
+			model.addAttribute("questionDetails", service.findQuestion(idx));
+		}
+
+		return "qna/show";
+	}
+
+}
