+++ src/main/java/kr/codesqaud/cafe/question/dto/response/QuestionWriteDTO.java
@@ -1,0 +1,33 @@
+package kr.codesqaud.cafe.question.dto.response;
+
+import java.time.LocalDateTime;
+
+import kr.codesqaud.cafe.question.domain.Question;
+
+public class QuestionWriteDTO {
+	private final String writer;
+	private final String title;
+	private final String contents;
+
+	public QuestionWriteDTO(String writer, String title, String contents) {
+		this.writer = writer;
+		this.title = title;
+		this.contents = contents;
+	}
+
+	public String getWriter() {
+		return writer;
+	}
+
+	public String getTitle() {
+		return title;
+	}
+
+	public String getContents() {
+		return contents;
+	}
+
+	public Question toEntity(int idx) {
+		return new Question(idx, writer, title, contents, LocalDateTime.now());
+	}
+}
