+++ src/main/java/kr/codesqaud/cafe/user/repository/H2UserRepository.java
@@ -1,0 +1,41 @@
+package kr.codesqaud.cafe.user.repository;
+
+import java.util.List;
+import java.util.NoSuchElementException;
+
+import org.springframework.context.annotation.Primary;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Repository;
+
+import kr.codesqaud.cafe.user.dto.request.SignUpDTO;
+import kr.codesqaud.cafe.user.dto.response.UserDTO;
+
+@Repository
+@Primary
+public class H2UserRepository implements UserRepository {
+	private final JdbcTemplate jdbcTemplate;
+
+	public H2UserRepository(JdbcTemplate jdbcTemplate) {
+		this.jdbcTemplate = jdbcTemplate;
+	}
+
+	public void insert(SignUpDTO dto) {
+		String sql = "INSERT INTO \"user\"(userId, password, name, email) VALUES (?, ?, ?, ?)";
+		jdbcTemplate.update(sql, dto.getUserId(), dto.getPassword(), dto.getName(), dto.getEmail());
+	}
+
+	@Override
+	public List<UserDTO> selectAll() {
+		return null;
+	}
+
+	@Override
+	public UserDTO selectByUserId(String userId) throws NoSuchElementException {
+		return null;
+	}
+
+	@Override
+	public void update(SignUpDTO dto) throws NoSuchElementException {
+		// TODO document why this method is empty
+	}
+}
